sell page and other minor bugs fixed
