+++ src/Components/News.jsx
@@ -1,41 +1,68 @@
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-import NewsList from './NewsList';
-import Pagination from './Pagination';
-import Loader from '../Utilis/Loader/Loader';
+import React, { useState, useEffect } from "react";
+import { useNavigate } from "react-router-dom";
 
-const ArticleList = () => {
-  const [articles, setArticles] = useState([]);
-  const [currentPage,setCurrentPage]=useState(1);
-  const [postPerPage,setPostPerPage]=useState(3);
+function Header() {
+  const [isMobile, setIsMobile] = useState(window.innerWidth < 768);
+  const navigate = useNavigate();
+
+  const handleResize = () => {
+    setIsMobile(window.innerWidth < 768);
+  };
 
   useEffect(() => {
-    axios.get('https://newsapi.org/v2/everything?q=finance&from=2024-06-13&sortBy=publishedAt&apiKey=e255d757ed5a44e7a64b2e646abe5fca')
-      .then(response => {
-        setArticles(response.data.articles);
-      })
-      .catch(error => {
-        console.error('Error fetching data:', error);
-      });
+    window.addEventListener("resize", handleResize);
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
   }, []);
-  const lastPostIndex=currentPage*postPerPage;
-  const firstPostIndex=lastPostIndex-postPerPage;
-  const currentPosts=articles.slice(firstPostIndex,lastPostIndex);
-  return (
-    <div className="container mx-auto mt-8">
-      <h1 className="text-5xl flex items-center text-[#52057B] justify-center font-bold mb-4">Latest Articles</h1>
-      <div className="flex-col">
-        <center>{articles.length==0 && <Loader />}</center>
-        <NewsList  articles={currentPosts} />
-        <Pagination
-                totalPosts={articles.length<40?articles.length:40}
-                postsPerPage={postPerPage}
-                setCurrentPage={setCurrentPage}
-                currentPage={currentPage}
-            />
+
+  const handleSubmitProfile = () => {
+    navigate("/dash-board");
+  };
+
+  const renderMenuItems = () => (
+    <>
+      <div className="cursor-pointer hover:underline bg-[#C3ACD0] font-bold text-[#52057B] px-4 py-2 rounded-md">
+        Buy
       </div>
+      <div className="cursor-pointer hover:underline bg-[#C3ACD0] font-bold text-[#52057B] px-4 py-2 rounded-md">
+        Sell
+      </div>
+      <div
+        className="cursor-pointer hover:underline bg-[#C3ACD0] font-bold text-[#52057B] px-4 py-2 rounded-md"
+        onClick={handleSubmitProfile}
+      >
+        Profile
+      </div>
+      <div className="cursor-pointer hover:underline bg-[#C3ACD0] font-bold text-[#52057B] px-4 py-2 rounded-md">
+        Logout
+      </div>
+    </>
+  );
+
+  return (
+    <div>
+      {isMobile ? (
+        <div className="fixed inset-y-0 left-0 w-64 bg-white shadow-md p-4 flex flex-col">
+          <div className="text-4xl mb-6 font-bold text-[#52057B]">
+            StockInv
+          </div>
+          <div className="flex flex-col gap-4">
+            {renderMenuItems()}
+          </div>
+        </div>
+      ) : (
+        <div className=" p-4 bg-[#FFFBF5] flex justify-between items-center">
+          <div className="text-6xl ml-4 font-bold text-[#52057B]">
+            StockInv
+          </div>
+          <div className="flex gap-4 items-center">
+            {renderMenuItems()}
+          </div>
+        </div>
+      )}
     </div>
   );
-};
+}
 
-export default ArticleList;
+export default Header;
